Move common variables to common structures
