+++ CMakeLists.txt
@@ -250,6 +250,7 @@ set(PICOHTTP_LIBRARY_FILES
     picohttp/h3zero_server.c
     picohttp/h3zero_uri.c
     picohttp/h3zero_url_template.c
+    picohttp/h09_server.h
     picohttp/picomask.c
     picohttp/quicperf.c
     picohttp/webtransport.c
@@ -260,6 +261,7 @@ set(PICOHTTP_HEADERS
      picohttp/h3zero_common.h
      picohttp/h3zero_uri.h
      picohttp/h3zero_url_template.h
+     picohttp/h09_server.h
      picohttp/democlient.h
      picohttp/demoserver.h
      picohttp/pico_webtransport.h
