+++ picohttp/picohttp.vcxproj
@@ -139,6 +139,8 @@
   <ItemGroup>
     <ClCompile Include="democlient.c" />
     <ClCompile Include="demoserver.c" />
+    <ClCompile Include="h09_common.c" />
+    <ClCompile Include="h09_server.c" />
     <ClCompile Include="h3zero.c" />
     <ClCompile Include="h3zero_client.c" />
     <ClCompile Include="h3zero_common.c" />
@@ -153,6 +155,8 @@
   <ItemGroup>
     <ClInclude Include="democlient.h" />
     <ClInclude Include="demoserver.h" />
+    <ClInclude Include="h09_common.h" />
+    <ClInclude Include="h09_server.h" />
     <ClInclude Include="h3zero.h" />
     <ClInclude Include="h3zero_common.h" />
     <ClInclude Include="h3zero_uri.h" />
